Add HSLtoRGB conversion
Processing doesn't support the HSB colorspace, so in order to do HSL 
operations & target certain colors / lightnesses, some HSL functions 
must be implemented. The toRGB() method takes an HSL color and converts 
it into RGB. The lightness() method takes a RGB color and gets the 
lightness of that color (this will be useful in later slider apply() 
methods).