Add SharpnessSlider
Added SharpnessSlider.pde, & Kernel.pde. Kernel is used for sharpness 
convolution and pSharp is used to store the sharpness convolution. 
Convolution occurs the frame after edit has been loaded. SharpnessSlider 
sharpens the image if the handle is pulled to the right or blurs the 
image if the handle is pulled to the left.